+++ spec/jahuty/data/render_spec.rb
@@ -2,13 +2,7 @@ module Jahuty
   module Data
     RSpec.describe Render do
       describe ".from" do
-        let(:data) { { id: 1, content: "foo" } }
-
-        it "raises error if :id key does not exist" do
-          data.delete(:id)
-
-          expect{ Render.from(data) }.to raise_error(ArgumentError)
-        end
+        let(:data) { { content: "foo" } }
 
         it "raises error if :content key does not exist" do
           data.delete(:content)
@@ -19,7 +13,7 @@ module Jahuty
 
       describe "#to_s" do
         it "returns content" do
-          expect(Render.new(1, "foo").to_s).to eq("foo")
+          expect(Render.new("foo").to_s).to eq("foo")
         end
       end
     end